refactor(domain/lotto): relocate Lotto class to match subject rule
- Moved Lotto class directly under package 'lotto' instead of 'lotto.domain'.
- Adjusted structure according to subject rule guideline.
