+++ src/main/kotlin/lotto/domain/LottoEvaluator.kt
@@ -1,5 +1,7 @@
 package lotto.domain
 
+import lotto.Lotto
+
 object LottoEvaluator {
 	fun evaluate(
 		lottoList: List<Lotto>,
