+++ src/main/kotlin/lotto/service/GameService.kt
@@ -1,6 +1,6 @@
 package lotto.service
 
-import lotto.domain.Lotto
+import lotto.Lotto
 
 import lotto.domain.LottoIssuer
 import lotto.policy.GamePolicy
